backend/requirements.txt
anyio==3.3.4
apturl==0.5.2
arrow==1.2.1
asgiref==3.4.1
bcrypt==3.1.7
binaryornot==0.4.4
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.3
command-not-found==0.3
cookiecutter==1.7.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.1.0
duplicity==0.8.12.0
email-validator==1.1.3
entrypoints==0.3
fastapi==0.70.0
fasteners==0.14.1
future==0.18.2
h11==0.12.0
httplib2==0.14.0
httptools==0.2.0
idna==2.8
Jinja2==3.0.2
jinja2-time==0.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
manage-fastapi==1.1.1
MarkupSafe==2.0.1
monotonic==1.5
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
poyo==0.5.0
prompt-toolkit==3.0.21
protobuf==3.6.1
pycairo==1.16.2
pycups==1.9.73
pydantic==1.8.2
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.19.1
python-slugify==5.0.2
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
questionary==1.10.0
reportlab==3.5.34
requests==2.26.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sniffio==1.2.0
starlette==0.16.0
svn=1.0.1
systemd-python==234
text-unidecode==1.3
typer==0.4.0
typing-extensions==3.10.0.2
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
uvicorn==0.15.0
uvloop==0.16.0
wadllib==1.3.3
watchgod==0.7
wcwidth==0.2.5
websockets==10.0
xkit==0.0.0
